refactor: move predicate variant to modulo core
