+++ source/modulo_core/include/modulo_core/PredicateVariant.hpp
@@ -2,8 +2,11 @@
 
 #include <variant>
 
-namespace modulo_utils {
+namespace modulo_core {
 
+/**
+ * @typedef PredicateVariant.
+ */
 typedef std::variant<bool, std::function<bool(void)>> PredicateVariant;
 
 }// namespace modulo_components::modulo_utils
